requirements.txt
streamlit
langchain
groq
chromadb
sqlite-utils
python-dotenv
tiktoken
